requirement.txt
apt-xapian-index==0.47
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
Click==7.0
command-not-found==0.3
cryptography==2.3.1
cupshelpers==1.0
distro-info==0.18
Django==2.1
django-javascript-settings==1.1.2
django-rest-framework==0.1.0
djangorestframework==3.8.2
Flask==1.0.2
Flask-SocketIO==3.0.2
gunicorn==19.9.0
httplib2==0.9.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.0.8
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==1.1.0
numpy==1.15.1
olefile==0.45.1
pandas==0.23.4
pexpect==4.2.1
Pillow==5.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pubnub==4.1.0
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycryptodomex==3.6.6
pycups==1.9.73
pygobject==3.26.1
PyGreSQL==5.0.6
PyJWT==1.6.4
PyMySQL==0.9.2
pyOpenSSL==18.0.0
python-apt==1.6.2
python-dateutil==2.7.3
python-debian==0.1.32
python-engineio==2.3.0
python-socketio==2.0.0
pytz==2018.5
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scipy==1.1.0
scour==0.36
SecretStorage==2.3.1
six==1.11.0
ssh-import-id==5.7
systemd-python==234
tornado==3.2.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
virtualenv==16.0.0
Werkzeug==0.14.1
xkit==0.0.0
